+++ internal/js_parser/js_parser_lower_class.go
@@ -376,8 +376,16 @@ func (p *parser) maybeLowerSuperPropertyGetInsideCall(call *js_ast.ECall) {
 	call.Args = append([]js_ast.Expr{thisExpr}, call.Args...)
 }
 
+type lowerAllInstanceFields uint8
+
+const (
+	lowerAllInstanceFields_false lowerAllInstanceFields = iota
+	lowerAllInstanceFields_true_skipSuperCallShim
+	lowerAllInstanceFields_true_needSuperCallShim
+)
+
 type classLoweringInfo struct {
-	lowerAllInstanceFields bool
+	lowerAllInstanceFields lowerAllInstanceFields
 	lowerAllStaticFields   bool
 	shimSuperCtorCalls     bool
 }
@@ -456,7 +464,7 @@ func (p *parser) computeClassLoweringInfo(class *js_ast.Class) (result classLowe
 				}
 			} else {
 				if p.privateSymbolNeedsToBeLowered(private) {
-					result.lowerAllInstanceFields = true
+					result.lowerAllInstanceFields = lowerAllInstanceFields_true_needSuperCallShim
 
 					// We can't transform this:
 					//
@@ -495,7 +503,7 @@ func (p *parser) computeClassLoweringInfo(class *js_ast.Class) (result classLowe
 				}
 			} else {
 				if p.options.unsupportedJSFeatures.Has(compat.ClassPrivateField) {
-					result.lowerAllInstanceFields = true
+					result.lowerAllInstanceFields = lowerAllInstanceFields_true_needSuperCallShim
 					result.lowerAllStaticFields = true
 				}
 			}
@@ -562,17 +570,22 @@ func (p *parser) computeClassLoweringInfo(class *js_ast.Class) (result classLowe
 				result.lowerAllStaticFields = true
 			}
 		} else {
-			// Instance fields must be lowered if the target doesn't support them
-			if p.options.unsupportedJSFeatures.Has(compat.ClassField) {
-				result.lowerAllInstanceFields = true
-			}
-
-			// Convert instance fields to assignment statements if the TypeScript
-			// setting for this is enabled. I don't think this matters for private
-			// fields because there's no way for this to call a setter in the base
-			// class, so this isn't done for private fields.
 			if p.options.ts.Parse && !class.UseDefineForClassFields {
-				result.lowerAllInstanceFields = true
+				// Convert instance fields to assignment statements if the TypeScript
+				// setting for this is enabled. I don't think this matters for private
+				// fields because there's no way for this to call a setter in the base
+				// class, so this isn't done for private fields.
+				if prop.InitializerOrNil.Data != nil {
+					result.lowerAllInstanceFields = lowerAllInstanceFields_true_needSuperCallShim
+				} else if result.lowerAllInstanceFields != lowerAllInstanceFields_true_needSuperCallShim {
+					// We can skip the "super()" call shim if all instance fields
+					// disappear completely when lowered. This happens when
+					// "useDefineForClassFields" is false and there is no initializer.
+					result.lowerAllInstanceFields = lowerAllInstanceFields_true_skipSuperCallShim
+				}
+			} else if p.options.unsupportedJSFeatures.Has(compat.ClassField) {
+				// Instance fields must be lowered if the target doesn't support them
+				result.lowerAllInstanceFields = lowerAllInstanceFields_true_needSuperCallShim
 			}
 		}
 	}
@@ -580,7 +593,7 @@ func (p *parser) computeClassLoweringInfo(class *js_ast.Class) (result classLowe
 	// We need to shim "super()" inside the constructor if this is a derived
 	// class and there are any instance fields that need to be lowered, since
 	// those use "this" and we can only access "this" after "super()" is called
-	if result.lowerAllInstanceFields && class.ExtendsOrNil.Data != nil {
+	if result.lowerAllInstanceFields == lowerAllInstanceFields_true_needSuperCallShim && class.ExtendsOrNil.Data != nil {
 		result.shimSuperCtorCalls = true
 	}
 
@@ -1046,7 +1059,7 @@ func (p *parser) lowerClass(stmt js_ast.Stmt, expr js_ast.Expr, result visitClas
 			if prop.Flags.Has(js_ast.PropertyIsStatic) {
 				mustLowerField = classLoweringInfo.lowerAllStaticFields
 			} else {
-				mustLowerField = classLoweringInfo.lowerAllInstanceFields
+				mustLowerField = classLoweringInfo.lowerAllInstanceFields != lowerAllInstanceFields_false
 			}
 		}
 
@@ -1784,17 +1797,28 @@ func (p *parser) insertStmtsAfterSuperCall(body *js_ast.FnBody, stmtsToInsert []
 	//   var __super = (...args) => {
 	//     super(...args);
 	//     ...stmtsToInsert...
+	//     return this;
 	//   };
 	//
 	argsRef := p.newSymbol(ast.SymbolOther, "args")
 	p.currentScope.Generated = append(p.currentScope.Generated, argsRef)
-	stmtsToInsert = append([]js_ast.Stmt{{Loc: body.Loc, Data: &js_ast.SExpr{Value: js_ast.Expr{Loc: body.Loc, Data: &js_ast.ECall{
+	p.recordUsage(argsRef)
+	superCall := js_ast.Expr{Loc: body.Loc, Data: &js_ast.ECall{
 		Target: js_ast.Expr{Loc: body.Loc, Data: js_ast.ESuperShared},
 		Args:   []js_ast.Expr{{Loc: body.Loc, Data: &js_ast.ESpread{Value: js_ast.Expr{Loc: body.Loc, Data: &js_ast.EIdentifier{Ref: argsRef}}}}},
-	}}}}}, stmtsToInsert...)
+	}}
+	stmtsToInsert = append(append(
+		[]js_ast.Stmt{{Loc: body.Loc, Data: &js_ast.SExpr{Value: superCall}}},
+		stmtsToInsert...),
+		js_ast.Stmt{Loc: body.Loc, Data: &js_ast.SReturn{ValueOrNil: js_ast.Expr{Loc: body.Loc, Data: js_ast.EThisShared}}},
+	)
+	if p.options.minifySyntax {
+		stmtsToInsert = p.mangleStmts(stmtsToInsert, stmtsFnBody)
+	}
 	body.Block.Stmts = append([]js_ast.Stmt{{Loc: body.Loc, Data: &js_ast.SLocal{Decls: []js_ast.Decl{{
 		Binding: js_ast.Binding{Loc: body.Loc, Data: &js_ast.BIdentifier{Ref: superCtorRef}}, ValueOrNil: js_ast.Expr{Loc: body.Loc, Data: &js_ast.EArrow{
 			HasRestArg: true,
+			PreferExpr: true,
 			Args:       []js_ast.Arg{{Binding: js_ast.Binding{Loc: body.Loc, Data: &js_ast.BIdentifier{Ref: argsRef}}}},
 			Body:       js_ast.FnBody{Loc: body.Loc, Block: js_ast.SBlock{Stmts: stmtsToInsert}},
 		}},